refactor: pages 내 디렉토리 이름에 page 삭제
