+++ src/pages/cardRegistrationComplete/CardRegistrationCompletePage.styled.ts
