+++ src/pages/newCard/NewCardPage.styled.ts
